backend/requirements.txt
Django==3.2.3
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
Pillow==9.0.0
python-dotenv==1.0.1
PyYAML==6.0
flake8==7.2.0